etw-genai-monitor/requirements.txt
psutil
google-generativeai
jsonschema
python-dateutil
fastapi
uvicorn[standard]
google-genai
numpy
